+++ pramen/core/src/test/resources/log4j.properties
@@ -19,10 +19,11 @@ log4j.appender.console.layout.ConversionPattern=%d{yy/MM/dd HH:mm:ss} %p %c{1}: 
 log4j.appender.console.Threshold=ERROR
 
 # Suppress warnings logged within various components
-log4j.logger.za.co.absa.pramen.core.utils.ConfigUtils$=OFF
-log4j.logger.za.co.absa.pramen.core.utils.JdbcNativeUtils$=OFF
+log4j.logger.za.co.absa.pramen.core.mocks.DummyProcessRunner=OFF
 log4j.logger.za.co.absa.pramen.core.notify.SyncNotificationEmail=OFF
+log4j.logger.za.co.absa.pramen.core.reader.TableReaderJdbc=OFF
 log4j.logger.za.co.absa.pramen.core.runner.task.TaskRunnerBase=OFF
 log4j.logger.za.co.absa.pramen.core.runner.task.TaskRunnerParallel=OFF
-log4j.logger.za.co.absa.pramen.core.reader.TableReaderJdbc=OFF
-log4j.logger.za.co.absa.pramen.core.mocks.DummyProcessRunner=OFF
+log4j.logger.za.co.absa.pramen.core.state.PipelineStateImpl=OFF
+log4j.logger.za.co.absa.pramen.core.utils.ConfigUtils$=OFF
+log4j.logger.za.co.absa.pramen.core.utils.JdbcNativeUtils$=OFF
