+++ pramen/examples/enceladus_single_config/daily_ingestion.conf
@@ -93,7 +93,7 @@ pramen.sources = [
 pramen.sinks = [
   {
     name = "dce"
-    factory.class = "za.co.absa.pramen.builtin.sink.EnceladusSink"
+    factory.class = "za.co.absa.pramen.extras.sink.EnceladusSink"
 
     format = "csv"
