+++ pramen/examples/enceladus_sourcing/sinks.conf
@@ -14,7 +14,7 @@
 pramen.sinks = [
   {
     name = "dce"
-    factory.class = "za.co.absa.pramen.builtin.sink.EnceladusSink"
+    factory.class = "za.co.absa.pramen.extras.sink.EnceladusSink"
 
     format = "csv"
