+++ pramen/extras/pom.xml
@@ -111,7 +111,7 @@
                     <transformers>
                         <transformer implementation="org.apache.maven.plugins.shade.resource.ServicesResourceTransformer" />
                         <transformer implementation="org.apache.maven.plugins.shade.resource.ManifestResourceTransformer">
-                            <mainClass>za.co.absa.pramen.runner.PipelineRunner</mainClass>
+                            <mainClass>za.co.absa.pramen.core.runner.PipelineRunner</mainClass>
                         </transformer>
                         <transformer implementation="org.apache.maven.plugins.shade.resource.AppendingTransformer">
                             <resource>reference.conf</resource>
