+++ pramen/extras/src/main/scala/za/co/absa/pramen/extras/writer/TableWriterKafka.scala
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package za.co.absa.pramen.writer
+package za.co.absa.pramen.extras.writer
 
 import com.typesafe.config.Config
 import org.apache.spark.sql.functions.struct
@@ -24,9 +24,9 @@ import za.co.absa.abris.avro.functions.to_avro
 import za.co.absa.abris.avro.read.confluent.SchemaManagerFactory
 import za.co.absa.abris.config.{AbrisConfig, ToAvroConfig}
 import za.co.absa.pramen.api.TableWriter
-import za.co.absa.pramen.avro.AvroUtils.{convertSparkToAvroSchema, fixNullableFields}
+import za.co.absa.pramen.extras.avro.AvroUtils.{convertSparkToAvroSchema, fixNullableFields}
 import za.co.absa.pramen.core.utils.ConfigUtils
-import za.co.absa.pramen.writer.model.{KafkaWriterConfig, NamingStrategy}
+import za.co.absa.pramen.extras.writer.model.{KafkaWriterConfig, NamingStrategy}
 
 import java.time.LocalDate
