+++ pramen/extras/src/test/scala/za/co/absa/pramen/tests/avro/AvroUtilsSuite.scala
@@ -19,7 +19,7 @@ package za.co.absa.pramen.tests.avro
 import org.apache.spark.sql.functions.struct
 import org.scalatest.WordSpec
 import za.co.absa.pramen.NestedDataFrameFactory
-import za.co.absa.pramen.avro.AvroUtils
+import za.co.absa.pramen.extras.avro.AvroUtils
 import za.co.absa.pramen.base.SparkTestBase
 import za.co.absa.pramen.fixtures.TextComparisonFixture
 import za.co.absa.pramen.core.utils.ResourceUtils.getResourceString
